Remove DBunit related test files and dependencies
A significant part of test data handling setup, including the associated DBunit-related test files and dependencies have been removed. This change was prompted by a shift toward utilizing pure SQL files for setting up test data in rental tables. This approach proved to be simpler and more in line with the rest of the codebase, reducing the complexity of the application and remaining consistent with the existing data handling techniques.